Move hetu validation from service to annotation. Fix controller tests to use valid (test) hetus.
