+++ oppijanumerorekisteri-api/src/main/java/fi/vm/sade/oppijanumerorekisteri/dto/HenkiloHetuAndOidDto.java
@@ -1,5 +1,6 @@
 package fi.vm.sade.oppijanumerorekisteri.dto;
 
+import fi.vm.sade.oppijanumerorekisteri.validation.ValidateHetu;
 import lombok.*;
 
 import java.io.Serializable;
@@ -11,7 +12,12 @@ import java.util.Date;
 @AllArgsConstructor
 @NoArgsConstructor
 public class HenkiloHetuAndOidDto implements Serializable {
+    private static final long serialVersionUID = 813852020644110186L;
+
     private String oidhenkilo;
+
+    @ValidateHetu
     private String hetu;
+
     private Date vtjsynced;
 }
